Adicionando múltiplos inimigos
